Move PropGridPanel event handlers into their own file
